backend/requirements.txt
# FastAPI und Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Bildverarbeitung
Pillow==10.1.0

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Pydantic für Datenvalidierung
pydantic==2.5.0

# Entwicklung und Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Production
gunicorn==21.2.0

# Logging und Monitoring
structlog==23.2.0
